split predictor.
